agent-sandbox-local/sandbox_cli/pyproject.toml
[project]
name = "sandbox-cli"
version = "2.0.0"
description = "Multi-backend Sandbox CLI - OrbStack & Tmux support"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "docker>=7.0.0",
    "libtmux>=0.37.0",
]

[project.scripts]
sbx = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]
